// ClientApp/src/app/spares-management/models/Procurement.ts
export interface Procurement {
    procurementId: number;
    demandId: number;
    demandTypeId:number;
    itemDetailId: number;
    itemCategoryId:number;
    procurementStatusId: number;
    principalNameId: number;
    manufactureId:number;
    localAgentId: number;
    supplierId: number;
    supplierAId: number;
    supplierMId: number;
    cstTecId: number;
    partOfShipmentId: number;
    departmentNameId: number;
    departmentName:string;
    itemDetail:string;
    itemName:string;
    demandType:string;
    cstTec:string;
    supplier:string;
    sparesCategoryId: number;
    verificationCompletStatus: number;
    tenderNumber: string;
    workOrderDate:Date;
    dateOfTenderFloat: Date;
    tenderopeningDate: Date;
    tenderPublishDate: Date;
    tenderNotice: string;
    tenderSpecification: string;
    financialApproval: string;
    workOrder: string;
    dateOfDelivery: Date;
    unitPrice: string;
    qty: string;
    remarks: string;
    procurementDocument: string;
    status: number;
    menuPosition: number;
    isActive: boolean;   
}