// ClientApp/src/app/spares-management/models/Acceptance.ts
export interface Acceptance {
    acceptanceId: number;
    procurementId: number;
    demandId: number;
    demandTypeId:number;
    itemDetailId:number;
    itemCategoryId:number;
    departmentNameId: number;
    departmentName:string;
    itemDetail:string;
    itemName:string;
    demandType:string;
    condition:string;
    procurementStatusId: number;
    sourceOfSupplyId:number;
    manufactureId:number;
    principalNameId: number;
    placeOfDeliveryId: number;
    demandAuthorityId:number;
    conditionOfItemId: number;
    sparesCategoryId: number;
    itemInspectionId:number;
    sftDate: Date;
    sftLetterNo:string;
    workOrderNo: string;
    workOrderDate: Date;
    sftQty:number;
    qty:number;
    storeQtyStatus:number;
    verificationCompletStatus:number;
    storeQty:number;
    //procurementQty:number;
    itemSerNo:string;
    model:string;
    brand: string;
    warranty: string;
    warrantyFrom:Date;
    warrantyTo:Date;
    deliveryDate:Date;
    inspectionDate: Date;
    purchasePrice:string;
    dateOfManufacture: Date;
    acDocument: string;
    arcDocument:string;
    cofcDocument: string;
    othersDocument: string;
    sftRegPage:string;
    //doc:string;
    acceptanceDocument:string;
    docVerification: string;
    sftStatus: boolean;
    remarks:string;
    menuPosition:string;
    isActive: boolean;
}